6-71-72-Authorization Deleting a User's Task

+++ nestjs-task-management/src/tasks/tasks.service.ts
@@ -21,7 +21,7 @@ export class TasksService {
   }
 
   async getTaskById(id: number, user: User): Promise<Task> {
-    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
+    const found = await this.taskRepository.findOne({ where: { id, userId: user.id } }); //where is needed
     if (!found) {
       throw new NotFoundException(`Task with ID "${id}" not found`);
     }
@@ -33,8 +33,8 @@ export class TasksService {
   }
 
   //Tutorial solution to challenge 4-47
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id }); // no where needed
     //console.log(result);
     if (result.affected === 0) {
       //if there were no rows affected
